Moved nullable check code into a separated class
